db_guardian/pyproject.toml
[project]
name = "db-guardian"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.0",
    "cryptography>=46.0.3",
    "django>=4.2,<5.0",
    "django-celery-beat>=2.8.1",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.1",
    "django-redis>=5.4.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-nested-routers>=0.95.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "pymysql>=1.1.2",
    "python-decouple>=3.8",
    "redis>=7.1.0",
    "sqlparse>=0.5.4",
    "whitenoise>=6.11.0",
]
